Shell: Create a Matahari API
The API takes the form of a store for the session state, wrapped around the
core API. Currently, acessing objects is done synchronously, so the cache
must be deleted (and the objects re-fetched) to see updated parameter
values. In the future, I expect to change this to get parameter updates
asynchronously.