build: Don't autogenerate man pages, use copies in source directory temporarily.
This adds a new cmake macro to handle installing man pages if one exists
in the source directory and adds a man page for every binary. The change is
temporary so that man pages are installed while autogeneration is not working.

Once autogeneration of man pages is working again, all the CMakeLists
changes can be changed back with the exception of the broker agent which will
never be autogenerated (the new install_manpage macro could stay too).

Signed-off-by: Jeff Peeler <jpeeler@redhat.com>